Agrego alertas de todos los modales y pantallas

+++ src/components/screens/Departamento/Departamento.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState } from 'react';
 import { useAuthStore } from '../../../auth/store/authStore';
 import axios from 'axios';
 import { ModalDepartamento } from '../../ui/ModalDepartamento/ModalDepartamento';
+import Swal from 'sweetalert2';
 
 interface Departamento {
     id: number;
@@ -47,17 +48,42 @@ export const Departamento = () => {
   }
 
   const eliminarDepartamento= async (id: number) => {
-    try {
-        await axios.delete(`http://localhost:9000/api/departamentos/${id}`, {
-            headers: {
-            Authorization: `Bearer ${token}`,
-            },
+    const result = await Swal.fire({
+          title: '¿Estás seguro?',
+          text: 'Esta acción eliminará el beneficiario de forma permanente.',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#d33',
+          cancelButtonColor: '#3085d6',
+          confirmButtonText: 'Sí, eliminar',
+          cancelButtonText: 'Cancelar'
         });
-        setDepartamentos(prev => prev.filter(b => b.id !== id));
-        } catch (error) {
-        console.error('Error al eliminar Departamento:', error);
-    }
+    
+      if (result.isConfirmed) {
+        try {
+            await axios.delete(`http://localhost:9000/api/departamentos/${id}`, {
+                headers: {
+                Authorization: `Bearer ${token}`,
+                },
+            });
+            setDepartamentos(prev => prev.filter(b => b.id !== id));
+            Swal.fire({
+                      icon: 'success',
+                      title: 'Eliminado',
+                      text: 'El departamento fue eliminado correctamente.',
+                      timer: 1500,
+                      showConfirmButton: false
+                    });
+            } catch (error) {
+            console.error('Error al eliminar Departamento:', error);
+                    Swal.fire({
+                      icon: 'error',
+                      title: 'Error',
+                      text: 'No se pudo eliminar el departamento.',
+                    });
+        }
   }
+}
 
   const handleCloseModal = () => {
     setIsModalOpen(false);
