+++ src/components/screens/Empresa/Empresa.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState } from 'react';
 import { useAuthStore } from '../../../auth/store/authStore';
 import axios from 'axios';
 import { ModalEmpresa } from '../../ui/ModalEmpresa/ModalEmpresa';
+import Swal from 'sweetalert2';
 
 interface Empresa {
     id: number;
@@ -51,16 +52,41 @@ export const Empresa = () => {
   }
 
   const eliminarEmpresa= async (id: number) => {
-    try {
-        await axios.delete(`http://localhost:9000/api/empresas/${id}`, {
-            headers: {
-            Authorization: `Bearer ${token}`,
-            },
-        });
-        setEmpresas(prev => prev.filter(b => b.id !== id));
-        } catch (error) {
-        console.error('Error al eliminar Empresa:', error);
-    }
+    const result = await Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Esta acción eliminará la empresa de forma permanente.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar'
+    });
+
+    if (result.isConfirmed) {
+      try {
+          await axios.delete(`http://localhost:9000/api/empresas/${id}`, {
+              headers: {
+              Authorization: `Bearer ${token}`,
+              },
+          });
+          setEmpresas(prev => prev.filter(b => b.id !== id));
+                  Swal.fire({
+                    icon: 'success',
+                    title: 'Eliminado',
+                    text: 'La empresa fue eliminada correctamente.',
+                    timer: 1500,
+                    showConfirmButton: false
+                  });
+          } catch (error) {
+          console.error('Error al eliminar Empresa:', error);
+                  Swal.fire({
+                    icon: 'error',
+                    title: 'Error',
+                    text: 'No se pudo eliminar la empresa.',
+                  });
+      }
+  }
   }
 
   const handleCloseModal = () => {
