+++ src/components/screens/Nacionalidad/Nacionalidad.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState } from 'react';
 import { useAuthStore } from '../../../auth/store/authStore';
 import axios from 'axios';
 import { ModalNacionalidad } from '../../ui/ModalNacionalidad/ModalNacionalidad';
+import Swal from 'sweetalert2';
 
 interface Nacionalidad {
     id: number;
@@ -42,15 +43,40 @@ export const Nacionalidad = () => {
   }
 
   const eliminarNacionalidad = async (id: number) => {
-    try {
-        await axios.delete(`http://localhost:9000/api/nacionalidades/${id}`, {
-            headers: {
-            Authorization: `Bearer ${token}`,
-            },
-        });
-        setNacionalidades(prev => prev.filter(b => b.id !== id));
-        } catch (error) {
-        console.error('Error al eliminar beneficiario:', error);
+    const result = await Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Esta acción eliminará la nacionalidad de forma permanente.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar'
+    });
+
+    if (result.isConfirmed) {    
+      try {
+          await axios.delete(`http://localhost:9000/api/nacionalidades/${id}`, {
+              headers: {
+              Authorization: `Bearer ${token}`,
+              },
+          });
+          setNacionalidades(prev => prev.filter(b => b.id !== id));
+          Swal.fire({
+            icon: 'success',
+            title: 'Eliminado',
+            text: 'La nacionalidad fue eliminada correctamente.',
+            timer: 1500,
+            showConfirmButton: false
+          });        
+          } catch (error) {
+          console.error('Error al eliminar beneficiario:', error);
+          Swal.fire({
+          icon: 'error',
+          title: 'Error',
+          text: 'No se pudo eliminar la nacionalidad.',
+          });
+      }
     }
   }
