+++ src/components/screens/Roles/Roles.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState } from 'react';
 import { useAuthStore } from '../../../auth/store/authStore';
 import axios from 'axios';
 import { ModalRol } from '../../ui/ModalRol/ModalRol';
+import Swal from 'sweetalert2';
 
 interface Rol {
     id: number;
@@ -45,15 +46,40 @@ export const Roles = () => {
   }
 
   const eliminarRol= async (id: number) => {
-    try {
-        await axios.delete(`http://localhost:9000/api/roles/${id}`, {
-            headers: {
-            Authorization: `Bearer ${token}`,
-            },
-        });
-        setRoles(prev => prev.filter(b => b.id !== id));
-        } catch (error) {
-        console.error('Error al eliminar roles:', error);
+    const result = await Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Esta acción eliminará el rol de forma permanente.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar'
+    });
+
+    if (result.isConfirmed) {    
+      try {
+          await axios.delete(`http://localhost:9000/api/roles/${id}`, {
+              headers: {
+              Authorization: `Bearer ${token}`,
+              },
+          });
+          setRoles(prev => prev.filter(b => b.id !== id));
+          Swal.fire({
+            icon: 'success',
+            title: 'Eliminado',
+            text: 'El rol fue eliminado correctamente.',
+            timer: 1500,
+            showConfirmButton: false
+          });          
+          } catch (error) {
+          console.error('Error al eliminar roles:', error);
+          Swal.fire({
+            icon: 'error',
+            title: 'Error',
+            text: 'No se pudo eliminar el rol.',
+          });          
+      }
     }
   }
